Hello World! - Hello SDL
Lazy Foo' Productions to the rescue.
By the way, the close() function appears to not work.
It seems putting the SDL_FreeSurface and SDL_DestroyWindow or probably other SDL functions inclusive into a user-defined function can lead to a segementation fault.
The exact error message is like this: `Segmentation fault (core dumped)`